+++ app/src/main/java/com/farma/parkinsoftapp/presentation/doctor/new_patient_tests/NewPatientsTestViewModel.kt
@@ -36,7 +36,7 @@ class NewPatientsTestViewModel @Inject constructor(
 
     fun createPatient(
         navigationArgs: NewPatientTestsRoute
-    ): Int {
+    ): Long {
         return mainRepository.addNewPatient(navigationArgs.convertToPatient())
     }
